10.6.3-preview (Last Update: 2015-04-25)
    On some Windows versions negative time value in 'date-list' for dates
    before 1970-1-1 are not handled by gmtime() and could crash newLISP.
    Now an error 44 is thrown when gmtime() cannot handle negative values.

    Fixed 'det' and 'mat' for Ubuntu Linux when using wrong dimensions.
    Tested on OSX, OpenBSD, FreeBSD Ubuntu Linux, Windows.

    Changes in several files for compiling a 64-bit version for Windows
    using mingw64. Thanks to Shigeru Kobayashi for doing this work.

    Filenames win32xxx.c change to win-xxx.c.

    The 'dolist' function now also works on arrays.

    The 'net-eval' functions now return the result of the last expression
    in the string argument, not the first.
        (net-eval host port "(+ 3 4) (* 3 4)" => 12 ; previously 7
    In previous versions, although all expressions in 'net-eval' would be
    evaluated, the result of the first was returned. The possibility
    of multiple expressions in a string was not documented. The new
    behaviour is more in line with the behaviour of 'eval-string' and
    'load'.

    Fixed a bug on Windows when APACHE was starting newlisp and no
    NEWLISPDIR wnd PROGRAMFILES environment variables were defined.

    New makefiles for Windows:
	makefile_mingw64 makefile_mingw64_ffi
	makefile_mingw64_utf8 makefile_mingw64_utf8_ffi
	makefile_mingw64dll makefile_mingw64dll_ffi
	makefile_mingw64dll_utf8 makefile_mingw64dll_utf8_ffi
    are working. Tested using tdm gcc 64 (MinGW based).

    The configure utility now uses the mingw64 makefiles if LLP64 is present.
    The configure-alt utility still only compiles for 32-bit.

    Both, 32-bit and 64-bit newLISP on Windows, now have ostype "Windows".
    In modules canvas.lsp and gsl.lsp ostype has been changed to "Windows"
    from "Win32". Same change in many qa-* files.

    C type cleanup for string size_t length in some C-functions.

    Server mode -http did not reject net-eval requests.

    net-eval connection timeout changed to 15 secs from 60 secs.
    Processing timeout still set by user in net-eval parameter.

    Added HTTP_AUTHORIZATION to environment variables in HTTP server.

    Additional syntax for 'date-value' -> (date-value list-date-time).
    Additional syntax (date-list) now same as (date-list (date-value)).

    The trim function without trim-char parameters will trim all
    white-space characters, not only spaces.

+++ win-dll.c
@@ -1,4 +1,4 @@
-/* win32dll.c - make the newlisp.exe usable as a DLL 
+/* win-dll.c - make the newlisp.exe usable as a DLL 
 
     Copyright (C) 2011 Lutz Mueller
 
@@ -163,7 +163,7 @@ currently only tested with newLISP as caller:
 
 */
 
-long EXPORT newlispCallback(char * funcName, long funcAddr, char * callType)
+intptr_t EXPORT newlispCallback(char * funcName, intptr_t funcAddr, char * callType)
 {
 CELL * pCell;
 SYMBOL * symbol;
